Fix filter selecting only known nodes
The filter was missing a negation ("!"). Without it, it is querying all neighbours of the root node and is only true if the recommended node is contained in its neighbours. Thereby, all unkown nodes are discarded. As we specifically want the unkown nodes for the user, I added the negation.

Also, I know I introduce a side effect here, but we would elliminate querying the nodes of the root node for every recommended node. Also converting the list to a HashSet to have O(1) data access.